backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
websockets==12.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
kafka-python==2.0.2
pydantic==2.6.1
pydantic-settings==2.2.0
httpx==0.26.0
python-dotenv==1.0.1
asyncpg==0.29.0
aioredis==2.0.1
